#121 Move inc column somewhere more appropriate
