+++ extension/ts/main/adapters/sheets/index.ts
@@ -1,4 +1,4 @@
-import {googleFetch} from "../services/google/googleFetch";
+import {googleFetch} from "../../services/google/googleFetch";
 
 const BASE_URL = "https://sheets.googleapis.com/v4/spreadsheets";
